refactor: move mocks to separate file
